+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [Fix] Fixed bug that showed wrong error when querying snippet with invalid function (#902)
 * [Fix] Disabled CTE generation when snippets are detected in a non-SELECT type query. (#651, #652)
 * [Fix] Fix empty result in certain duckdb `SELECT` and `SUMMARIZE` queries with leading comments (#892)
+* [Doc] Added `run_statements` to the Python API docs (#922)
 
 ## 0.10.2 (2023-09-22)